Restructure pubsub api package
