+++ implementation/pubsub/src/main/java/org/inaetics/dronessimulator/pubsub/api/broker/Publisher.java
@@ -1,4 +1,6 @@
-package org.inaetics.dronessimulator.pubsub.api;
+package org.inaetics.dronessimulator.pubsub.api.broker;
+
+import org.inaetics.dronessimulator.pubsub.api.Message;
 
 import java.io.IOException;
