+++ implementation/pubsub/src/main/java/org/inaetics/dronessimulator/pubsub/impl/broker/rabbitmq/RabbitPublisher.java
@@ -2,9 +2,9 @@ package org.inaetics.dronessimulator.pubsub.impl.broker.rabbitmq;
 
 import com.rabbitmq.client.Connection;
 import org.inaetics.dronessimulator.pubsub.api.Message;
-import org.inaetics.dronessimulator.pubsub.api.Publisher;
-import org.inaetics.dronessimulator.pubsub.api.Serializer;
-import org.inaetics.dronessimulator.pubsub.api.Topic;
+import org.inaetics.dronessimulator.pubsub.api.broker.Publisher;
+import org.inaetics.dronessimulator.pubsub.api.serializer.Serializer;
+import org.inaetics.dronessimulator.pubsub.api.broker.Topic;
 
 import java.io.IOException;
