+++ implementation/pubsub/src/main/java/org/inaetics/dronessimulator/pubsub/impl/serializer/java/JavaSerializer.java
@@ -1,7 +1,7 @@
 package org.inaetics.dronessimulator.pubsub.impl.serializer.java;
 
 import org.inaetics.dronessimulator.pubsub.api.Message;
-import org.inaetics.dronessimulator.pubsub.api.Serializer;
+import org.inaetics.dronessimulator.pubsub.api.serializer.Serializer;
 
 import java.io.*;
